chore(ios): refactor code base structure
