+++ ios/IonRemote/ViewModels/SessionViewModel+Commands.swift
@@ -1,0 +1,247 @@
+import Foundation
+
+// MARK: - Commands
+
+extension SessionViewModel {
+
+    func sync() {
+        send(.sync)
+    }
+
+    func sendSync() {
+        send(.sync)
+    }
+
+    func sendPrompt(tabId: String, text: String) {
+        send(.prompt(tabId: tabId, text: text))
+    }
+
+    func cancel(tabId: String) {
+        send(.cancel(tabId: tabId))
+    }
+
+    func rewindConversation(tabId: String, messageId: String) {
+        send(.rewind(tabId: tabId, messageId: messageId))
+    }
+
+    func forkFromMessage(tabId: String, messageId: String) {
+        send(.forkFromMessage(tabId: tabId, messageId: messageId))
+    }
+
+    func respondPermission(tabId: String, questionId: String, optionId: String) {
+        send(.respondPermission(tabId: tabId, questionId: questionId, optionId: optionId))
+    }
+
+    /// Dismiss a special permission card (AskUserQuestion/ExitPlanMode) without
+    /// sending respond_permission -- the tool was already auto-allowed on desktop.
+    func dismissSpecialPermission(tabId: String, questionId: String) {
+        if let idx = tabs.firstIndex(where: { $0.id == tabId }) {
+            tabs[idx].permissionQueue.removeAll { $0.questionId == questionId }
+        }
+        if questionId.hasPrefix("restored-") {
+            dismissedRestoredCards.insert(questionId)
+        } else {
+            // Live card dismissed -- block restoredSpecialCard from re-triggering
+            dismissedLiveSpecialTabs.insert(tabId)
+        }
+    }
+
+    func loadConversation(tabId: String) {
+        guard !loadingConversation.contains(tabId) else { return }
+        messages.removeValue(forKey: tabId)
+        messageCountByTab.removeValue(forKey: tabId)
+        conversationLoaded.remove(tabId)
+        conversationHasMore.removeValue(forKey: tabId)
+        conversationCursor.removeValue(forKey: tabId)
+        conversationLoadFailed.remove(tabId)
+        loadingConversation.insert(tabId)
+        send(.loadConversation(tabId: tabId, before: nil))
+        startLoadTimer(tabId: tabId)
+    }
+
+    func clearConversation(tabId: String) {
+        messages.removeValue(forKey: tabId)
+        messageCountByTab.removeValue(forKey: tabId)
+        conversationLoaded.remove(tabId)
+        conversationHasMore.removeValue(forKey: tabId)
+        conversationCursor.removeValue(forKey: tabId)
+        dismissedRestoredCards = dismissedRestoredCards.filter { !$0.hasPrefix("restored-") }
+    }
+
+    func loadMoreMessages(tabId: String) {
+        guard !loadingConversation.contains(tabId),
+              conversationHasMore[tabId] == true,
+              let cursor = conversationCursor[tabId] else { return }
+        loadingConversation.insert(tabId)
+        send(.loadConversation(tabId: tabId, before: cursor))
+        startLoadTimer(tabId: tabId)
+    }
+
+    func startLoadTimer(tabId: String) {
+        conversationLoadTimers[tabId]?.cancel()
+        conversationLoadTimers[tabId] = Task { @MainActor [weak self] in
+            try? await Task.sleep(for: .seconds(15))
+            guard !Task.isCancelled, let self else { return }
+            guard self.loadingConversation.contains(tabId) else { return }
+            let retries = self.conversationLoadRetryCount[tabId] ?? 0
+            if retries < 1 {
+                // First timeout -- retry once
+                self.conversationLoadRetryCount[tabId] = retries + 1
+                let cursor = self.conversationCursor[tabId]
+                self.send(.loadConversation(tabId: tabId, before: cursor))
+                self.startLoadTimer(tabId: tabId)
+            } else {
+                // Second timeout -- give up
+                self.loadingConversation.remove(tabId)
+                self.conversationLoadFailed.insert(tabId)
+                self.conversationLoadTimers.removeValue(forKey: tabId)
+                self.conversationLoadRetryCount.removeValue(forKey: tabId)
+            }
+        }
+    }
+
+    func cancelLoadTimer(tabId: String) {
+        conversationLoadTimers[tabId]?.cancel()
+        conversationLoadTimers.removeValue(forKey: tabId)
+        conversationLoadRetryCount.removeValue(forKey: tabId)
+    }
+
+    func createTab(workingDirectory: String? = nil) {
+        let dir = workingDirectory ?? defaultBaseDirectory
+        send(.createTab(workingDirectory: dir))
+    }
+
+    func closeTab(_ tabId: String) {
+        pendingCloseTabIds.insert(tabId)
+        send(.closeTab(tabId: tabId))
+        tabs.removeAll { $0.id == tabId }
+        liveText.removeValue(forKey: tabId)
+        messages.removeValue(forKey: tabId)
+        loadingConversation.remove(tabId)
+        conversationLoaded.remove(tabId)
+        conversationHasMore.removeValue(forKey: tabId)
+        conversationCursor.removeValue(forKey: tabId)
+    }
+
+    func setPermissionMode(tabId: String, mode: PermissionMode) {
+        // Optimistic local update for responsive UI
+        if let idx = tabs.firstIndex(where: { $0.id == tabId }) {
+            tabs[idx].permissionMode = mode
+        }
+        send(.setPermissionMode(tabId: tabId, mode: mode))
+    }
+
+    // MARK: - Terminal Commands
+
+    func createTerminalTab(workingDirectory: String? = nil) {
+        let dir = workingDirectory ?? defaultBaseDirectory
+        send(.createTerminalTab(workingDirectory: dir))
+    }
+
+    // MARK: - Engine Commands
+
+    func createEngineTab(workingDirectory: String? = nil, profileId: String? = nil) {
+        let dir = workingDirectory ?? defaultBaseDirectory
+        send(.createEngineTab(workingDirectory: dir, profileId: profileId))
+    }
+
+    func submitEnginePrompt(tabId: String, text: String) {
+        let key = engineCompoundKey(tabId: tabId)
+        enginePinnedPrompt[key] = text
+        // Add user message to conversation
+        var msgs = engineMessages[key] ?? []
+        msgs.append(EngineMessage(id: UUID().uuidString, role: "user", content: text, timestamp: Date().timeIntervalSince1970 * 1000))
+        engineMessages[key] = msgs
+        // Set tab running
+        if let idx = tabs.firstIndex(where: { $0.id == tabId }) {
+            tabs[idx].status = .running
+        }
+        let instanceId = activeEngineInstance[tabId]
+        send(.enginePrompt(tabId: tabId, text: text, instanceId: instanceId))
+    }
+
+    func abortEngine(tabId: String) {
+        let instanceId = activeEngineInstance[tabId]
+        send(.engineAbort(tabId: tabId, instanceId: instanceId))
+    }
+
+    func respondEngineDialog(tabId: String, dialogId: String, value: String) {
+        let key = engineCompoundKey(tabId: tabId)
+        engineDialogs[key] = nil
+        let instanceId = activeEngineInstance[tabId]
+        send(.engineDialogResponse(tabId: tabId, dialogId: dialogId, value: value, instanceId: instanceId))
+    }
+
+    // MARK: - Engine Instance Commands
+
+    func addEngineInstance(tabId: String) {
+        send(.engineAddInstance(tabId: tabId))
+    }
+
+    func removeEngineInstance(tabId: String, instanceId: String) {
+        send(.engineRemoveInstance(tabId: tabId, instanceId: instanceId))
+    }
+
+    func selectEngineInstance(tabId: String, instanceId: String) {
+        activeEngineInstance[tabId] = instanceId
+        send(.engineSelectInstance(tabId: tabId, instanceId: instanceId))
+        // Load conversation for the newly selected instance
+        loadEngineConversation(tabId: tabId)
+    }
+
+    func loadEngineConversation(tabId: String) {
+        let instanceId = activeEngineInstance[tabId]
+        print("[Ion] loadEngineConversation: tabId=\(tabId), instanceId=\(instanceId ?? "nil"), instances=\(engineInstances[tabId]?.map(\.id) ?? [])")
+        send(.loadEngineConversation(tabId: tabId, instanceId: instanceId))
+    }
+
+    func sendTerminalInput(tabId: String, instanceId: String, data: String) {
+        send(.terminalInput(tabId: tabId, instanceId: instanceId, data: data))
+    }
+
+    func sendTerminalResize(tabId: String, instanceId: String, cols: Int, rows: Int) {
+        send(.terminalResize(tabId: tabId, instanceId: instanceId, cols: cols, rows: rows))
+    }
+
+    func addTerminalInstance(tabId: String) {
+        send(.terminalAddInstance(tabId: tabId))
+    }
+
+    func removeTerminalInstance(tabId: String, instanceId: String) {
+        send(.terminalRemoveInstance(tabId: tabId, instanceId: instanceId))
+    }
+
+    func selectTerminalInstance(tabId: String, instanceId: String) {
+        activeTerminalInstance[tabId] = instanceId
+        send(.terminalSelectInstance(tabId: tabId, instanceId: instanceId))
+    }
+
+    func requestTerminalSnapshot(tabId: String) {
+        send(.requestTerminalSnapshot(tabId: tabId))
+    }
+
+    func renameTab(tabId: String, customTitle: String?) {
+        if let idx = tabs.firstIndex(where: { $0.id == tabId }) {
+            tabs[idx].customTitle = customTitle
+        }
+        send(.renameTab(tabId: tabId, customTitle: customTitle))
+    }
+
+    func renameTerminalInstance(tabId: String, instanceId: String, label: String) {
+        terminalInstanceLabels["\(tabId):\(instanceId)"] = label
+        send(.renameTerminalInstance(tabId: tabId, instanceId: instanceId, label: label))
+    }
+
+    func terminalInstanceLabel(tabId: String, instanceId: String, fallback: String) -> String {
+        terminalInstanceLabels["\(tabId):\(instanceId)"] ?? fallback
+    }
+
+    // MARK: - Send
+
+    func send(_ command: RemoteCommand) {
+        guard let transport else { return }
+        Task {
+            try? await transport.send(command)
+        }
+    }
+}
